Implement exactly-once QoS publish protocol flow.
This commit also removes the previously-seen-packets set added in
b72afb57015ab593efb695c66c03beecfa11a3af and no longer ignores duplicate
PUBLISH packets. Initially the plan was to to clear the previously-seen flag
for a particular packet when a packet with a sufficiently higher identifier
was received. But this by itself is not reliable, and there's no easy way
to be sure the server has received a particular PUBACK.

Instead, there is now a new type for publications received from the server,
and this type exposes the `dup` flag. It is now the responsibility of *users*
of `Client` to identify whether a duplicate publication really is one they've
seen before or not.