Consider packet identifiers exhausted when a single reserved ID is encountered.
Before this change, the bitset would attempt to find the next available
packet identifier by skipping over reserved identifiers. However this is not
very performant, and would only be necessary if a packet identifier used
2^16 packets ago was somehow still reserved in the first place.

Since such a situation is very extraordinary, this change makes it so that
the bitset gives up immediately when this happens. This also matches the
behavior of libraries like mosquitto.