Return errors to the user for invalid user-provided packets instead of killing the client
The user now receives an error if:

- they attempt to subscribe or unsubscribe with a topic filter that cannot be
  encoded into a SUBSCRIBE / UNSUBSCRIBE packet, say because it is too large.

- they attempt to publish a publication that cannot be encoded into
  a PUBLISH packet, say because the topic filter or payload is too large.

Previously these malformed requests would be accepted, and the client would
later fail while trying to encode them. It would then reset the connection,
but continue to attempt to re-send them and fail again, repeatedly.

Another case that could happen is if the user subscribed to (or unsubscribed
from) a large number of topics at the same time. It was possible that while
each topic filter individually was not too big, the batched
SUBSCRIBE / UNSUBSCRIBE packet became too large to encode. The client now
knows how to split such (un)subscriptions into multiple packets.

Fixes #1